ai-backend/pyproject.toml
[project]
name = "ai-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "openrouter>=0.1.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
